Functioning Trivia Game with flowchart
This is a functional trivia game that works with text files in the same folder. The text file must be as so, and can be as many as desirable. 
"title
category
question add / for new line but keep question on 1 line
answer 1
answer 2
answer 3 
answer 4
correct answer
explanation
category
question add / for new line but keep question on 1 line
answer 1
answer 2
answer 3 
answer 4
correct answer
explanation"

The game made with this is titled: "test.txt" (Upload later/separate file)